first try

+++ lib/features/race/domain/race_repository.dart
@@ -5,17 +5,17 @@ import 'race.dart';
 abstract class RacesRepository {
   /// Récupère toutes les courses
   Future<List<Race>> getRaces();
-  
+
   /// Récupère les courses d'un raid spécifique
   Future<List<Race>> getRacesByRaidId(int raidId);
-  
+
   /// Récupère une course par son ID
   Future<Race?> getRaceById(int id);
   Future<int> getRegisteredTeamsCount(int raceId);
 
-  Future<int> createRace(Race race, Map<int, double> categoryPrices);
+  Future<int> createRace(Race race, Map<int, int> categoryPrices);
   Future<List<Category>> getCategories();
-  Future<Map<int, double>> getRaceCategoryPrices(int raceId);
+  Future<Map<int, int>> getRaceCategoryPrices(int raceId);
   Future<bool> canAddRaceToRaid(int raidId);
   Future<int> getRaceCount(int raidId);
   Future<int?> getMaxRaceCount(int raidId);
